+++ HISTORY.rst
@@ -2,6 +2,15 @@
 History
 =======
 
+Releases
+--------
+
+2019-06-06
+~~~~~~~~~~
+
+- functions to read and write IDF files
+    - issue #20
+
 0.1.0 (2018-10-15)
 ------------------